+++ training/models/training_duration.py
@@ -1,0 +1,59 @@
+# -*- coding: utf-8 -*-
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+# © 2015 Grupo ESOC Ingeniería de Servicios, S.L.U.
+
+from openerp import api, fields, models
+from . import exceptions
+
+
+class DurationType(models.Model):
+    """Types of the training actions' durations.
+
+    See docs for :class:`~.ActionType`.
+    """
+
+    _name = "training.duration_type"
+    _sql_constraints = [("unique_name",
+                         "UNIQUE(name)",
+                         "Name must be unique.")]
+
+    name = fields.Char(required=True, index=True, translate=True)
+    duration_ids = fields.One2many(
+        "training.duration",
+        "type_id",
+        "Expected hours of this type",
+        help="Expected hours of this type defined in training actions.")
+    action_type_ids = fields.Many2many(
+        "training.action_type",
+        string="Training action types",
+        help="Training action types that expect this hour type.")
+
+
+class Duration(models.Model):
+    _name = "training.duration"
+    _sql_constraints = [("training_vs_hours_unique",
+                         "UNIQUE(type_id, action_id)",
+                         "Cannot repeat the hour type in a training action.")]
+
+    duration = fields.Float(default=0, required=True)
+    type_id = fields.Many2one(
+        "training.duration_type",
+        "Type of hours",
+        required=True)
+    action_id = fields.Many2one(
+        "training.action",
+        "Training action",
+        required=True)
+
+    @api.multi
+    @api.constrains("type_id", "action_id")
+    def _check_right_duration_types(self):
+        """Check that the hour types are the right ones."""
+
+        expected_types = (self.action_id.type_id
+                          .expected_duration_type_ids)
+
+        if expected_types and self.type_id not in expected_types:
+            raise exceptions.WrongDurationType(
+                self.type_id,
+                expected_types)
